feat: inline RequestBody into DetailView
