+++ src/utils.tsx
@@ -1,3 +1,5 @@
+import { ParamItem } from './admin-panel/ParamsTable'
+
 const htmlEscapesMap: any = {
 	'&': '&amp;',
 	'<': '&lt;',
@@ -9,3 +11,13 @@ const htmlEscapesMap: any = {
 export const escapeHtml = (input: string): string => {
 	return Object.keys(htmlEscapesMap).reduce((acc, search) => acc.replaceAll(search, htmlEscapesMap[search]), input)
 }
+
+interface ParamsObject {
+	[key: string]: string
+}
+
+export const mapParamsToObject = (params: ParamItem[]): ParamsObject =>
+	params
+		.filter((i) => i.isChecked)
+		.filter((i) => i.key || i.value)
+		.reduce((acc, i) => ({ ...acc, [i.key]: i.value }), {})
